Prepare more files for update to pdfium::span<>
Update files which were missed in the last go-around. In particular,
avoid using dynamic_extent with two-arg subspan (deviation from std::
spec in base::span<> that eliminates branching).

Bug: 42271100
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/130993
Commit-Queue: Tom Sepez <tsepez@chromium.org>

+++ core/fpdfapi/font/cpdf_cidfont.cpp
@@ -350,7 +350,8 @@ wchar_t CPDF_CIDFont::GetUnicodeFromCharCode(uint32_t charcode) const {
   wchar_t unicode;
   size_t ret = FX_MultiByteToWideChar(
       kCharsetCodePages[static_cast<size_t>(cmap_->GetCoding())],
-      ByteStringView(pdfium::make_span(sequence).first(charsize)),
+      ByteStringView(
+          pdfium::make_span(sequence).first(static_cast<size_t>(charsize))),
       pdfium::span_from_ref(unicode));
   return ret == 1 ? unicode : 0;
 #else
